Converted spectrum analyzer hardware module to new structure, corrected spectrum analyzer demo notebooks, removed some unused import statements.
