+++ pylabnet/launchers/pylabnet_server.py
@@ -17,13 +17,10 @@ However, you can also call this directly, with command-line arguments:
         :param port: (int) port number
 """
 
-import sys
-import time
 import importlib
 
 from pylabnet.utils.helper_methods import parse_args, show_console, hide_console
 from pylabnet.utils.logging.logger import LogClient
-from pylabnet.launchers import servers
 
 
 def main():
