+++ pylabnet/network/client_server/external_gui.py
@@ -2,7 +2,6 @@ import pickle
 
 from pylabnet.network.core.service_base import ServiceBase
 from pylabnet.network.core.client_base import ClientBase
-from pylabnet.network.core.generic_server import GenericServer
 
 
 class Service(ServiceBase):
@@ -53,7 +52,7 @@ class Service(ServiceBase):
 
     def exposed_assign_container(self, container_widget, container_label):
         return self._module.assign_container(self, container_widget, container_label)
-    
+
     def exposed_set_curve_data(self, data_pickle, plot_label, curve_label, error_pickle=None):
         data = pickle.loads(data_pickle)
         error = pickle.loads(error_pickle)
@@ -88,7 +87,7 @@ class Service(ServiceBase):
 
     def exposed_get_text(self, label_label):
         return pickle.dumps(self._module.get_text(label_label))
-    
+
     def exposed_was_button_pressed(self, event_label):
         return self._module.was_button_pressed(event_label)
 
@@ -146,7 +145,7 @@ class Client(ClientBase):
 
     def assign_container(self, container_widget, container_label):
         return self._service.exposed_assign_container(container_widget, container_label)
-    
+
     def set_curve_data(self, data, plot_label, curve_label, error=None):
         data_pickle = pickle.dumps(data)
         error_pickle = pickle.dumps(error)
@@ -181,7 +180,7 @@ class Client(ClientBase):
 
     def get_text(self, label_label):
         return pickle.loads(self._service.exposed_get_text(label_label))
-    
+
     def was_button_pressed(self, event_label):
         return self._service.exposed_was_button_pressed(event_label)
